Vector layer polygon rendering.

+++ src/og/shaderProgram/simple.js
@@ -24,7 +24,7 @@ og.shaderProgram.simple = function () {
                 vColor = aVertexColor;\
             }',
         fragmentShader:
-            'precision mediump float;\
+            'precision highp float;\
             varying vec4 vColor;\
             void main(void) {\
                 gl_FragColor = vColor;\
